refactor(#8): put use common tools and tool collection for all agent ✨
