+++ issue-solver/src/issue_solver/agents/tools/anthropic_tools_schema.py
